NPlusOneDBSpansDetector
